🚚 (#236): Move Mail record to commons
