+++ backend/app.hopps.commons/src/main/java/app/hopps/commons/mail/Mail.java
@@ -1,16 +1,10 @@
-package app.hopps.mailservice;
-
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.ObjectMapper;
+package app.hopps.commons.mail;
 
 import java.util.Arrays;
 import java.util.Map;
 import java.util.Objects;
 
 public record Mail(String[] mailReceivers, MailTemplates templateId, Map<String, String> variables) {
-
-    private static final ObjectMapper objectMapper = new ObjectMapper();
-
     @Override
     public boolean equals(Object o) {
         if (o == null || getClass() != o.getClass())
@@ -27,14 +21,10 @@ public record Mail(String[] mailReceivers, MailTemplates templateId, Map<String,
 
     @Override
     public String toString() {
-        try {
-            return objectMapper.writeValueAsString(this);
-        } catch (JsonProcessingException ignored) {
-            return "{" +
-                    "mailReceivers: [" + String.join(",", mailReceivers) + "]," +
-                    "templateId: " + templateId + ',' +
-                    "variables: " + variables +
-                    '}';
-        }
+        return "Mail{" +
+                "mailReceivers=" + Arrays.toString(mailReceivers) +
+                ", templateId=" + templateId +
+                ", variables=" + variables +
+                '}';
     }
 }
