+++ backend/app.hopps.commons/src/test/java/app/hopps/commons/mail/MailTest.java
@@ -1,4 +1,4 @@
-package app.hopps.mailservice;
+package app.hopps.commons.mail;
 
 import nl.jqno.equalsverifier.EqualsVerifier;
 import org.junit.jupiter.api.Test;
@@ -17,7 +17,7 @@ class MailTest {
     void testToString() {
         Mail mail = new Mail(new String[] { "info@hopps.de", "support@hopps.de" }, MailTemplates.TEMP, Map.of());
 
-        String expected = "{\"mailReceivers\":[\"info@hopps.de\",\"support@hopps.de\"],\"templateId\":\"TEMP\",\"variables\":{}}";
+        String expected = "Mail{mailReceivers=[info@hopps.de, support@hopps.de], templateId=TEMP, variables={}}";
 
         String actual = mail.toString();
